executor/requirements.txt
fastapi==0.132.0
uvicorn==0.41.0
docker==7.1.0
aiosqlite==0.22.1
pydantic==2.12.5
python-multipart==0.0.22
slowapi==0.1.9
PyJWT==2.11.0
